Windows 64 bit crash
gcc and clang treat long as 8 bytes on Linux and Mac OS X.  However
Visual Studio treats long as 4 bytes.  To make the software function the
same on windows as it does on the *nix platforms, we convert the long
types to exact width integer types in stdint.h

+++ lib/label/rectangle.c
@@ -137,7 +137,7 @@ unsigned int RectArea(Rect_t * r)
      */
     area = 1;
     for (i = 0; i < NUMDIMS; i++) {
-      long long a_test = area * r->boundary[i + NUMDIMS] - r->boundary[i];
+      int64_t a_test = area * r->boundary[i + NUMDIMS] - r->boundary[i];
       if( a_test > UINT_MAX) {
 	agerr (AGERR, "label: area too large for rtree\n");
 	return UINT_MAX;
